requirements.txt
amqp==5.0.2
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
constantly==15.1.0
Django==3.1.3
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.12.2
gevent==20.9.0
greenlet==0.4.17
grequests==0.6.0
h11==0.11.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
prompt-toolkit==3.0.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-threads==0.1.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
urllib3==1.26.2
uvicorn==0.12.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.2.0
